Rename attribute lib files.
